requirements.txt
Flask==2.3.2
requests
python-dotenv
Bootstrap_Flask==2.2.0
WTForms==3.0.1
Flask_WTF==1.2.1
Werkzeug==3.0.0
flask-ckeditor==0.4.10
Flask_SQLAlchemy==3.0.3
SQLAlchemy==2.0.19